init ResultsForm
